Add support for JSON-RPC v2
This adds a `version` argument to `Jsonrpc.string_of_{call;response}`,
and updates the `of_string` functions accordingly.

The main changes compared to JSON-RPC v1 are:
1. And added version field: `{jsonrpc: "2.0"}`.
2. Support for named parameters in requests:
   `params: {"name1": "value2", "name2", "value2"}`.
3. Responses have a `result` OR an `error`, but not both.

Regarding (2), since the `RPC.call` type defines its `params` as a `t list`, it
does not have support for named parameters in the type itself. Positional
params are still treated as in V1, by putting them in an array (`params:
[value1, value2]`). To overcome this, we assume that if the first parameter in
this list has an `Rpc.Dict` type, and this is the only parameter, that the
dictionary contains named parameters.

Fixes #32

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>